Move runTask to common evals package
