+++ apps/workers-bindings/evals/accounts.eval.ts
@@ -1,10 +1,11 @@
 import { expect } from 'vitest'
 import { describeEval } from 'vitest-evals'
 
+import { runTask } from '@repo/eval-tools/src/runTask'
 import { checkFactuality } from '@repo/eval-tools/src/scorers'
 import { eachModel } from '@repo/eval-tools/src/test-models'
 
-import { initializeClient, runTask } from './utils' // Assuming utils.ts will exist here
+import { initializeClient } from './utils' // Assuming utils.ts will exist here
 
 // Define a mock account ID for testing
 const MOCK_ACCOUNT_ID = 'mock-account-12345'
