+++ apps/workers-bindings/evals/hyperdrive.eval.ts
@@ -1,11 +1,12 @@
 import { expect } from 'vitest'
 import { describeEval } from 'vitest-evals'
 
+import { runTask } from '@repo/eval-tools/src/runTask'
 import { checkFactuality } from '@repo/eval-tools/src/scorers'
 import { eachModel } from '@repo/eval-tools/src/test-models'
 import { HYPERDRIVE_TOOLS } from '@repo/mcp-common/src/tools/hyperdrive'
 
-import { initializeClient, runTask } from './utils' // Assuming utils.ts will exist here
+import { initializeClient } from './utils' // Assuming utils.ts will exist here
 
 // TODO: Add test for creating hyperdrive config with the following params once we can securely pass parameters to the tool. See: https://github.com/modelcontextprotocol/modelcontextprotocol/pull/382
 // const HYPERDRIVE_NAME = 'neon-test-hyperdrive'
