+++ apps/workers-bindings/evals/kv_namespaces.eval.ts
@@ -1,11 +1,12 @@
 import { expect } from 'vitest'
 import { describeEval } from 'vitest-evals'
 
+import { runTask } from '@repo/eval-tools/src/runTask'
 import { checkFactuality } from '@repo/eval-tools/src/scorers'
 import { eachModel } from '@repo/eval-tools/src/test-models'
 import { KV_NAMESPACE_TOOLS } from '@repo/mcp-common/src/tools/kv_namespace'
 
-import { initializeClient, runTask } from './utils' // Assuming utils.ts will exist here
+import { initializeClient } from './utils' // Assuming utils.ts will exist here
 
 eachModel('$modelName', ({ model }) => {
 	describeEval('Create Cloudflare KV Namespace', {
