+++ packages/eval-tools/package.json
@@ -13,6 +13,7 @@
 	"dependencies": {
 		"@ai-sdk/openai": "1.3.20",
 		"@cloudflare/vitest-pool-workers": "0.8.14",
+		"agents": "0.0.67",
 		"ai": "4.3.10",
 		"workers-ai-provider": "0.3.0",
 		"wrangler": "4.10.0",
